Updated LinearProgressBar class for proper Cocoapod access
